+++ wand/image.py
@@ -60,6 +60,28 @@ class Image(object):
     :param filename: opens an image of the ``filename``
     :type filename: :class:`basestring`
 
+    .. describe:: [left:right, top:bottom]
+
+       Crops the image by its ``left``, ``right``, ``top`` and ``bottom``,
+       and then returns the cropped one. ::
+
+           with img[100:200, 150:300] as cropped:
+               # manipulated the cropped image
+               pass
+
+       Like other subscriptable objects, default is 0 or its width/height::
+
+           img[:, :]        #--> just clone
+           img[:100, 200:]  #--> equivalent to img[0:100, 200:img.height]
+
+       Negative integers count from the end (width/height)::
+
+           img[-70:-50, -20:-10]
+           #--> equivalent to img[width-70:width-50, height-20:height-10]
+
+       :returns: the cropped image
+       :rtype: :class:`Image`
+
     """
 
     __slots__ = '_wand',
@@ -177,6 +199,55 @@ class Image(object):
 
         """
         return type(self)(image=self)
+
+    def __getitem__(self, idx):
+        if isinstance(idx, collections.Iterable):
+            idx = tuple(idx)
+            d = len(idx)
+            if not (1 <= d <= 2):
+                raise ValueError('index cannot be {0}-dimensional'.format(d))
+            elif d == 2:
+                x, y = idx
+                x_slice = isinstance(x, slice)
+                y_slice = isinstance(y, slice)
+                if x_slice and not y_slice:
+                    y = slice(y, y + 1)
+                elif not x_slice and y_slice:
+                    x = slice(x, x + 1)
+                elif not (x_slice or y_slice):
+                    raise NotImplementedError('pixel getter is not '
+                                              'implemented yet')
+                if not (x.step is None and y.step is None):
+                    raise ValueError('slicing with step is unsupported')
+                elif (x.start is None and x.stop is None and
+                      y.start is None and y.stop is None):
+                    return self.clone()
+                def abs_(n, m, null=None):
+                    if n is None:
+                        return m if null is None else null
+                    elif not isinstance(n, numbers.Integral):
+                        raise TypeError('expected integer, not ' + repr(n))
+                    elif n > m:
+                        raise IndexError(repr(n) + ' > ' + repr(m))
+                    return m + n if n < 0 else n
+                x_start = abs_(x.start, self.width, 0)
+                x_stop = abs_(x.stop, self.width)
+                y_start = abs_(y.start, self.height, 0)
+                y_stop = abs_(y.stop, self.height)
+                width = x_stop - x_start
+                height = y_stop - y_start
+                if width < 1:
+                    raise ValueError('image width cannot be zero')
+                elif height < 1:
+                    raise ValueError('image width cannot be zero')
+                cloned = self.clone()
+                library.MagickCropImage(cloned.wand, width, height,
+                                        x_start, y_start)
+                self.raise_exception()
+                return cloned
+            else:
+                raise NotImplementedError('1d indexing is not implemented yet')
+        raise NotImplementedError('row getter is not implemented yet')
 
     @property
     def width(self):